+++ src/translations/translations_3.rs
@@ -9,6 +9,7 @@ use crate::{Language, StyleType};
 pub fn general_translation(language: Language) -> &'static str {
     match language {
         Language::EN | Language::RO => "General",
+        Language::CS => "Hlavní",
         // Language::FA => "عمومی",
         Language::ES => "Generales",
         Language::IT => "Generali",
@@ -44,6 +45,7 @@ pub fn zoom_translation(language: Language) -> &'static str {
         | Language::PT
         | Language::NL
         | Language::SV => "Zoom",
+        Language::CS => "Přiblížení",
         // Language::FA => "بزرگنمایی",
         Language::PL => "Powiększenie",
         Language::RU => "Масштаб интерфейса",
@@ -64,6 +66,7 @@ pub fn zoom_translation(language: Language) -> &'static str {
 pub fn mmdb_files_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Database files",
+        Language::CS => "Databázové soubory",
         // Language::FA => "پرونده های پایگاه داده",
         Language::ES => "Archivos de la base de datos",
         Language::IT => "File di database",
@@ -92,6 +95,7 @@ pub fn mmdb_files_translation(language: Language) -> &'static str {
 pub fn params_not_editable_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "The following parameters can't be modified during the analysis",
+        Language::CS => "Následující parametry nelze během analýzy měnit",
         // Language::FA => "مولفه های روبرو هنگام تحلیل قابل تغییر نیستند",
         Language::ES => "Los siguientes parámetros no pueden modificarse durante el análisis",
         Language::IT => "I seguenti parametri non sono modificabili durante l'analisi",
@@ -122,6 +126,7 @@ pub fn params_not_editable_translation(language: Language) -> &'static str {
 pub fn custom_style_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Custom style",
+        Language::CS => "Vlastní styl",
         // Language::FA => "شیوه سفارشی",
         Language::ES | Language::PT => "Estilo personalizado",
         Language::IT => "Stile personalizzato",
@@ -149,6 +154,7 @@ pub fn custom_style_translation(language: Language) -> &'static str {
 pub fn copy_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Copy",
+        Language::CS => "Kopírovat",
         // Language::FA => "رونوشت",
         Language::IT | Language::ES => "Copia",
         Language::FR | Language::RO => "Copie",
@@ -175,6 +181,7 @@ pub fn copy_translation(language: Language) -> &'static str {
 pub fn port_translation(language: Language) -> &'static str {
     match language {
         Language::EN
+        | Language::CS
         | Language::FR
         | Language::DE
         | Language::PL
@@ -202,6 +209,7 @@ pub fn port_translation(language: Language) -> &'static str {
 // pub fn invalid_filters_translation(language: Language) -> &'static str {
 //     match language {
 //         Language::EN => "Invalid filters",
+//         Language::CS => "Neplatné filtry",
 //         // Language::FA => "صافی نامعتبر",
 //         Language::ES | Language::PT => "Filtros inválidos",
 //         Language::IT => "Filtri non validi",
@@ -229,6 +237,7 @@ pub fn port_translation(language: Language) -> &'static str {
 pub fn messages_translation(language: Language) -> &'static str {
     match language {
         Language::EN | Language::FR => "Messages",
+        Language::CS => "Zprávy",
         // Language::FA => "پیام ها",
         Language::ES => "Mensajes",
         Language::IT => "Messaggi",
@@ -256,6 +265,7 @@ pub fn messages_translation(language: Language) -> &'static str {
 pub fn link_type_translation(language: Language) -> &'static str {
     match language {
         Language::EN | Language::NL => "Link type",
+        Language::CS => "Typ linky",
         // Language::FA => "نوع پیوند",
         Language::ES => "Tipo de conexión",
         Language::IT => "Tipo di collegamento",
@@ -287,6 +297,9 @@ pub fn unsupported_link_type_translation<'a>(
     let translation = match language {
         Language::EN => {
             "The link type associated with this adapter is not supported by Sniffnet yet..."
+        }
+        Language::CS => {
+            "Typ linky přidružený k tomuto adaptéru zatím není Sniffnetem podporován...."
         }
         // Language::FA => "نوع پیوند مرتبط با این مبدل هنوز توسط Sniffnet پشتیبانی نمی شود...",
         Language::ES => {
@@ -351,6 +364,7 @@ pub fn unsupported_link_type_translation<'a>(
 pub fn style_from_file_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Select style from a file",
+        Language::CS => "Výběr stylu ze souboru",
         // Language::FA => "انتخاب شیوه از یک پرونده",
         Language::ES => "Selecciona el estilo desde un archivo",
         Language::IT => "Seleziona lo stile da un file",
@@ -379,6 +393,7 @@ pub fn style_from_file_translation(language: Language) -> &'static str {
 pub fn database_from_file_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Select database file",
+        Language::CS => "Výběr souboru databáze",
         // Language::FA => "پرونده پایگاه داده را انتخاب کنید",
         Language::ES => "Selecciona un archivo de base de datos",
         Language::IT => "Seleziona file di database",
@@ -407,6 +422,7 @@ pub fn database_from_file_translation(language: Language) -> &'static str {
 pub fn filter_by_host_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Filter by network host",
+        Language::CS => "Filtrovat podle síťového hostitele",
         // Language::FA => "صافی بر اساس میزبان شبکه",
         Language::ES => "Filtra por host de red",
         Language::IT => "Filtra per host di rete",
@@ -435,6 +451,7 @@ pub fn filter_by_host_translation(language: Language) -> &'static str {
 pub fn service_translation(language: Language) -> &'static str {
     match language {
         Language::EN | Language::FR | Language::DE | Language::SV => "Service",
+        Language::CS => "Služba",
         // Language::FA => "خدمت",
         Language::ES => "Servicio",
         Language::IT => "Servizio",
@@ -460,6 +477,7 @@ pub fn service_translation(language: Language) -> &'static str {
 pub fn export_capture_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Export capture file",
+        Language::CS => "Export souboru zachycení",
         // Language::FA => "خروجی گرفتن پرونده تسخیری",
         Language::IT => "Esporta file di cattura",
         Language::FR => "Exporter le fichier de capture",
@@ -489,6 +507,7 @@ pub fn export_capture_translation(language: Language) -> &'static str {
 pub fn directory_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Directory",
+        Language::CS => "Složka",
         // Language::FA => "پوشه",
         Language::IT => "Cartella",
         Language::FR => "Répertoire",
@@ -515,6 +534,7 @@ pub fn directory_translation(language: Language) -> &'static str {
 pub fn select_directory_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Select destination directory",
+        Language::CS => "Výběr cílové složky",
         // Language::FA => "انتخاب پوشه مقصد",
         Language::IT => "Seleziona cartella di destinazione",
         Language::FR => "Sélectionner le répertoire de destination",
@@ -543,6 +563,7 @@ pub fn select_directory_translation(language: Language) -> &'static str {
 pub fn file_name_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "File name",
+        Language::CS => "Jméno souboru",
         // Language::FA => "نام پرونده",
         Language::IT => "Nome del file",
         Language::FR => "Nom du fichier",
@@ -571,6 +592,7 @@ pub fn file_name_translation(language: Language) -> &'static str {
 pub fn thumbnail_mode_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Thumbnail mode",
+        Language::CS => "Režim miniatury",
         // Language::FA => "حالت تصویر بندانگشتی",
         Language::IT => "Modalità miniatura",
         Language::FR => "Mode miniature",
@@ -598,6 +620,7 @@ pub fn thumbnail_mode_translation(language: Language) -> &'static str {
 // pub fn learn_more_translation(language: Language) -> &'static str {
 //     match language {
 //         Language::EN => "Do you want to learn more?",
+//         Language::CS => "Chcete se dozvědět více?",
 //         // Language::FA => "آیا می خواهید بیشتر یاد بگیرید؟",
 //         Language::IT => "Vuoi saperne di più?",
 //         Language::FR => "Voulez-vous en savoir davantage?",