+++ src/translations/types/language.rs
@@ -6,8 +6,8 @@ use serde::{Deserialize, Serialize};
 
 use crate::StyleType;
 use crate::countries::flags_pictures::{
-    CN, DE, ES, FI, FLAGS_WIDTH_BIG, FR, GB, GR, ID, IT, JP, KR, NL, PL, PT, RO, RU, SE, TR, TW,
-    UA, UZ, VN,
+    CN, CZ, DE, ES, FI, FLAGS_WIDTH_BIG, FR, GB, GR, ID, IT, JP, KR, NL, PL, PT, RO, RU, SE, TR,
+    TW, UA, UZ, VN,
 };
 
 /// This enum defines the available languages.
@@ -61,11 +61,14 @@ pub enum Language {
     ID,
     /// Dutch
     NL,
+    /// Czech
+    CS,
 }
 
 impl Language {
-    pub const ALL: [Language; 22] = [
+    pub const ALL: [Language; 23] = [
         Language::EN,
+        Language::CS,
         Language::DE,
         Language::EL,
         Language::ES,
@@ -114,6 +117,7 @@ impl Language {
             Language::VI => VN,
             Language::ID => ID,
             Language::NL => NL,
+            Language::CS => CZ,
         })))
         .width(FLAGS_WIDTH_BIG)
     }
@@ -134,6 +138,7 @@ impl Language {
                 | Language::ZH_TW
                 | Language::ID
                 | Language::ES
+                | Language::CS,
         )
     }
 }
@@ -164,6 +169,7 @@ impl fmt::Display for Language {
             Language::VI => "Tiếng Việt",
             Language::ID => "Bahasa Indonesia",
             Language::NL => "Nederlands",
+            Language::CS => "Čeština",
         };
         write!(f, "{self:?} - {lang_str}")
     }