+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All Sniffnet releases with the relative changes are documented in this file.
 
 ## [UNRELEASED]
+- Added Czech translation 🇨🇿 ([#960](https://github.com/GyulyVGC/sniffnet/pull/960))
 - Improve update checks using `semver` ([#891](https://github.com/GyulyVGC/sniffnet/pull/891))
 - Updated Spanish translation to v1.4 ([#955](https://github.com/GyulyVGC/sniffnet/pull/955))
 - Fix AppImage error with `libpcap` dependency ([#971](https://github.com/GyulyVGC/sniffnet/pull/971))